+++ front/lib/resources/content_fragment_resource.ts
@@ -1,5 +1,6 @@
 import type { ContentFragmentType, ModelId, Result } from "@dust-tt/types";
 import { Err, Ok } from "@dust-tt/types";
+import { Storage } from "@google-cloud/storage";
 import type {
   Attributes,
   CreationAttributes,
@@ -7,6 +8,7 @@ import type {
   Transaction,
 } from "sequelize";
 
+import appConfig from "@app/lib/api/config";
 import { Message } from "@app/lib/models";
 import { BaseResource } from "@app/lib/resources/base_resource";
 import { gcsConfig } from "@app/lib/resources/storage/config";
@@ -126,4 +128,39 @@ export function contentFragmentUrl({
     filePath,
     fileUrl: `https://storage.googleapis.com/${gcsConfig.getGcsPrivateUploadsBucket()}/${filePath}`,
   };
+}
+
+export async function storeContentFragmentText({
+  workspaceId,
+  conversationId,
+  messageId,
+  content,
+}: {
+  workspaceId: string;
+  conversationId: string;
+  messageId: string;
+  content: string;
+}): Promise<string | null> {
+  if (content === "") {
+    return null;
+  }
+
+  const { filePath, fileUrl } = contentFragmentUrl({
+    workspaceId,
+    conversationId,
+    messageId,
+    contentFormat: "text",
+  });
+  const storage = new Storage({
+    keyFilename: appConfig.getServiceAccount(),
+  });
+
+  const bucket = storage.bucket(gcsConfig.getGcsPrivateUploadsBucket());
+  const gcsFile = bucket.file(filePath);
+
+  await gcsFile.save(content, {
+    contentType: "text/plain",
+  });
+
+  return fileUrl;
 }
